LPS-203720 portal-search-web: Update tests
