--- modules/apps/portal-search/portal-search-web/src/test/java/com/liferay/portal/search/web/internal/facet/display/context/AssetVocabularyCategoriesSearchFacetDisplayContextTest.java
@@ -1,29 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.web.internal.facet.display.context;
-
-import com.liferay.portal.test.rule.LiferayUnitTestRule;
-
-import org.junit.ClassRule;
-import org.junit.Rule;
-
-/**
- * @author Bryan Engler
- */
-public class AssetVocabularyCategoriesSearchFacetDisplayContextTest
-	extends BaseCategoriesSearchFacetDisplayContextTestCase {
-
-	@ClassRule
-	@Rule
-	public static final LiferayUnitTestRule liferayUnitTestRule =
-		LiferayUnitTestRule.INSTANCE;
-
-	@Override
-	protected String getFacetFieldName() {
-		return "assetVocabularyCategoryIds";
-	}
-
-}
